requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
pyodbc==5.1.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
Flask-Migrate==4.0.5
Flask-Talisman==1.1.0
Pydantic==2.7.4
Flask-Limiter==3.8.0
PyJWT==2.9.0
langchain-google-genai
langchain
textblob
SpeechRecognition
gTTS
pydub
chromadb
sentence-transformers
